// src/providers/sentry/gen/zod/getOrganizationIssueSchema.ts
export const getOrganizationIssue200Schema = z.object({
    "isUnhandled": z.optional(z.boolean()),
"count": z.optional(z.string()),
"userCount": z.optional(z.int()),
"firstSeen": z.iso.datetime().nullish(),
"lastSeen": z.iso.datetime().nullish(),
"id": z.string(),
"shareId": z.nullable(z.string()),
"shortId": z.string(),
"title": z.string(),
"culprit": z.nullable(z.string()),
"permalink": z.string(),
"logger": z.nullable(z.string()),
"level": z.enum(["sample", "debug", "info", "warning", "error", "fatal", "unknown"]),
"status": z.enum(["resolved", "ignored", "pending_deletion", "pending_merge", "reprocessing", "unresolved"]),
"statusDetails": z.object({
    "autoResolved": z.optional(z.boolean()),
"ignoreCount": z.optional(z.int()),
"ignoreUntil": z.optional(z.iso.datetime()),
"ignoreUserCount": z.optional(z.int()),
"ignoreUserWindow": z.optional(z.int()),
"ignoreWindow": z.optional(z.int()),
"actor": z.optional(z.object({
    "identities": z.optional(z.array(z.object({
    "id": z.string(),
"name": z.string(),
"organization": z.object({
    "slug": z.string(),
"name": z.string()
    }),
"provider": z.object({
    "id": z.string(),
"name": z.string()
    }),
"dateVerified": z.iso.datetime(),
"dateSynced": z.iso.datetime()
    }))),
"avatar": z.optional(z.object({
    "avatarType": z.optional(z.string()),
"avatarUuid": z.string().nullish(),
"avatarUrl": z.string().nullish()
    })),
"authenticators": z.optional(z.array(z.any())),
"canReset2fa": z.optional(z.boolean()),
"id": z.string(),
"name": z.string(),
"username": z.string(),
"email": z.string(),
"avatarUrl": z.string(),
"isActive": z.boolean(),
"isSuspended": z.boolean(),
"hasPasswordAuth": z.boolean(),
"isManaged": z.boolean(),
"dateJoined": z.iso.datetime(),
"lastLogin": z.nullable(z.iso.datetime()),
"has2fa": z.boolean(),
"lastActive": z.nullable(z.iso.datetime()),
"isSuperuser": z.boolean(),
"isStaff": z.boolean(),
"experiments": z.object({
    
    }).catchall(z.any()),
"emails": z.array(z.object({
    "id": z.string(),
"email": z.string(),
"is_verified": z.boolean()
    }))
    })),
"inNextRelease": z.optional(z.boolean()),
"inRelease": z.optional(z.string()),
"inCommit": z.optional(z.string()),
"pendingEvents": z.optional(z.int()),
"info": z.optional(z.any())
    }),
"substatus": z.nullable(z.enum(["archived_until_escalating", "archived_until_condition_met", "archived_forever", "escalating", "ongoing", "regressed", "new"])),
"isPublic": z.boolean(),
"platform": z.nullable(z.string()),
"priority": z.nullable(z.enum(["low", "medium", "high"])),
"priorityLockedAt": z.nullable(z.iso.datetime()),
"seerFixabilityScore": z.nullable(z.number()),
"seerAutofixLastTriggered": z.nullable(z.iso.datetime()),
"seerExplorerAutofixLastTriggered": z.nullable(z.iso.datetime()),
"project": z.object({
    "id": z.string(),
"name": z.string(),
"slug": z.string(),
"platform": z.nullable(z.string())
    }),
"type": z.enum(["default", "error", "csp", "nel", "hpkp", "expectct", "expectstaple", "transaction", "generic", "feedback"]),
"issueType": z.string(),
"issueCategory": z.string(),
"metadata": z.object({
    
    }).catchall(z.any()),
"numComments": z.int(),
"assignedTo": z.nullable(z.object({
    "type": z.enum(["user", "team"]),
"id": z.string(),
"name": z.string(),
"email": z.optional(z.string())
    })),
"isBookmarked": z.boolean(),
"isSubscribed": z.boolean(),
"subscriptionDetails": z.nullable(z.object({
    "disabled": z.optional(z.boolean()),
"reason": z.optional(z.string())
    })),
"hasSeen": z.boolean(),
"annotations": z.array(z.object({
    "displayName": z.string(),
"url": z.string()
    })),
"firstRelease": z.object({
    
    }).catchall(z.any()).nullish(),
"lastRelease": z.object({
    
    }).catchall(z.any()).nullish(),
"tags": z.optional(z.array(z.object({
    
    }).catchall(z.any()))),
"stats": z.optional(z.object({
    
    }).catchall(z.array(z.array(z.number())))),
"inbox": z.object({
    "reason": z.int(),
"reason_details": z.nullable(z.object({
    "until": z.nullable(z.string()),
"count": z.nullable(z.int()),
"window": z.nullable(z.int()),
"user_count": z.nullable(z.int()),
"user_window": z.nullable(z.int())
    })),
"date_added": z.iso.datetime()
    }).nullish(),
"owners": z.array(z.object({
    "type": z.string(),
"owner": z.string(),
"date_added": z.iso.datetime()
    })).nullish(),
"forecast": z.optional(z.object({
    
    }).catchall(z.any())),
"integrationIssues": z.optional(z.array(z.object({
    
    }).catchall(z.any()))),
"sentryAppIssues": z.optional(z.array(z.object({
    "id": z.string(),
"issueId": z.string(),
"serviceType": z.string(),
"displayName": z.string(),
"webUrl": z.string()
    }))),
"latestEventHasAttachments": z.optional(z.boolean()),
"activity": z.array(z.object({
    
    }).catchall(z.any())),
"seenBy": z.array(z.object({
    
    }).catchall(z.any())),
"userReportCount": z.int(),
"participants": z.array(z.object({
    
    }).catchall(z.any()))
    }) as unknown as z.ZodType<GetOrganizationIssue200>
// ...
export const getOrganizationIssueQueryResponseSchema = z.lazy(() => getOrganizationIssue200Schema) as unknown as z.ZodType<GetOrganizationIssueQueryResponse>